Docker entry-point: decrypt-and-start
This is a port of the ApplauseOSS/kms-encryption-toolkit script,
`decrypt-and-start`, to Go. The intent is to have a small,
self-contained binary which can easily be using as an entry-point to
allow for decrypting strings made using the same methods as the
original. This is intended to be a drop-in replacement.

Signed-off-by: Chris Gianelloni <cgianelloni@applause.com>